+++ scrapper/src/main/java/backend/academy/linktracker/scrapper/service/TgChatService.java
@@ -1,5 +1,8 @@
 package backend.academy.linktracker.scrapper.service;
 
+import backend.academy.linktracker.scrapper.exceptions.ChatAlreadyRegisteredException;
+import backend.academy.linktracker.scrapper.exceptions.ChatNotFoundException;
+import backend.academy.linktracker.scrapper.properties.ScrapperMessages;
 import backend.academy.linktracker.scrapper.repository.ChatRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,14 +13,21 @@ import org.springframework.transaction.annotation.Transactional;
 public class TgChatService {
 
     private final ChatRepository chatRepository;
+    private final ScrapperMessages scrapperMessages;
 
     @Transactional
     public void registerChat(Long id) {
+        if (chatRepository.existsById(id)) {
+            throw new ChatAlreadyRegisteredException(scrapperMessages.getErrors().getChatAlreadyRegistered());
+        }
         chatRepository.save(id);
     }
 
     @Transactional
     public void deleteChat(Long id) {
+        if (!chatRepository.existsById(id)) {
+            throw new ChatNotFoundException(scrapperMessages.getErrors().getChatNotFound());
+        }
         chatRepository.deleteById(id);
     }
 }
